backend/requirements.txt
fastapi
psycopg2-binary
uvicorn
virtualenv
dotenv 
python-dotenv